switch between eth0 and eth1 roles.
It turns out that eth1 network device refuses to turn into promiscous
and pass received packets to the bridge.
This can be debuged inside the eth1 driver.
Instead a quicker approach of using eth0 instead is chosen.
eth0 to be an interface in the bridge with tap0.
eth1 to be used for ssh.